NO-SNOW: store VECTOR dimension as usize
Switch RowType::Vector's `dimension` from `i32` to `usize`. The
dimension is a count of elements and is non-negative by construction,
so `usize` mirrors the value space directly and keeps capacity math
(`capacity * dimension`) cast-free. The conversion friction moves to
the Arrow boundary (a single `as i32` in `arrow_utils` and
`FixedSizeListArray::try_new`), which matches how other Arrow-using
code handles size-typed conversions.

+++ sf_core/src/arrow_utils.rs
@@ -289,7 +289,7 @@ pub fn create_field_with_type(
             let data_type = data_type.unwrap_or_else(|| {
                 DataType::FixedSizeList(
                     Arc::new(Field::new_list_field(child_type, true)),
-                    *dimension,
+                    *dimension as i32,
                 )
             });
             Ok(Field::new(name, data_type, *nullable).with_metadata(metadata))
